python-llm/requirements.txt
# Database dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Environment and configuration
python-dotenv==1.0.0

# LLM and agents
openai==1.14.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
pydantic==2.5.3

# Web utilities
requests==2.31.0
beautifulsoup4==4.12.2

# Testing
pytest==7.4.3
